gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = ""
authors = ["Your Name <simon.garcial@udea.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.4"}
httpx = "^0.23.1"
aiohttp = {extras = ["async-timeout"], version = "^3.8.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
uvicorn = "^0.20.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
